Advise against crossing diode wires
Realised this when making my board, when I had ROW2COL/COL2ROW the wrong way
around in firmware, and half my keys didn't work (or so I thought -
actually none of them should have). The keys that worked were being
shorted by the crossed diodes.
This is a subtle mistake, because you won't notice it unless you're a
furious typer, but it defeats the point of having diodes and can lead to
ghosting when multiple keys are pressed at the same time.